misc updates

+++ simple-dop-example/src/gui_integration_mod/render.rs
@@ -45,7 +45,7 @@ impl<'a> RenderFn<CustomGuiData, GuiRenderingData<'a>> for GuiRenderFn {
 		
 		// text
 		if element.has_text {
-			let text_size = (element.text_size * element_area.height) as f32 * element_area.screen_size.1 as f32 * 0.7;
+			let text_size = element.text_size * element_area.height * element_area.screen_size.1 as f32 * 0.7;
 			let spacings = gui_utils::get_char_spacings(element, rendering_data.positioning_font, text_size);
 			let mut line_start_poss = Vec::with_capacity(element.text.len());
 			for (i, line) in element.text.iter().enumerate() {
